Save data in MongoDB
The data is capped at 50 KB per room, which is more than enough for a single
terminal screen (maybe even a bit overkill). Whenever a new person joins that
room, we get everything in the collection, put into a single dict, and send to
her.

There is a racing condition. The user might receive data out of order, with the
initial packet being received after some others, which might create some
weirdness. It should be rare, though.