Adds local caching for MongoDB
This is useful to avoid hitting the MongoDB's servers during every request to
check for Authorization keys and such. This gives a great speed boost.

In a synthetic benchmark using Apache's AB with 50 concurrent threads making
10.000 requests and using a local MongoDB instance, I got:

| Statistic | Without caching | With caching   |
| --------- | --------------- | -------------- |
| Req/s     | 776.67 (mean)   | 1106.35 (mean) |
| Time/req  | 64.378 (mean)   | 45.194 (mean)  |
| 99% Time  | 113 ms          | 74 ms          |

This test was the best-case, with all requests going to a single room with a
single authorization code. The difference will be smaller in production, but as
each person that's broadcasting POSTs data every second, it should still be a
good improvement. Specially considering that in production the database and web
servers are on different machines.