Delete MongoDB's collection when stopping to transmit.
This functionality isn't very important, but it made me refactor the code quite
a bit. While it's still messy, it's getting better :+1:

Note that only the room's history is deleted, and not its authorization. In
other words, if I created a room with an password (say to present something a
couple hours from now), it'll still be available. It'll only be deleted after
the TTL set in its collection expires (currently set at one day).