[#19] Exit shellshare when trying to send too much data (HTTP Error 413)
This isn't the ideal. For instance, we could simply split the data in smaller
chunks and tried POSTing those. This is the easiest, though, and we avoid
having to deal with people that try to transmit stuff like `cat /dev/urandom`.

The current limit, 300 KB, should be more than enough for most cases. If it's
not, we can raise it later.